feat(CustomRPC): adding toggle for default value
