+++ apps/davi/public/locales/en/translation.json
@@ -18,7 +18,10 @@
     "customRPC": "Custom RPC",
     "customRPCUrl": "Custom RPC URL",
     "enterURL": "Enter URL",
-    "saveChanges": "Save Changes"
+    "useDefaultValue": "use default value",
+    "saveChanges": "Save Changes",
+    "genericError": "Seems to be a problem with the url provided",
+    "wrongChain": "The url provided is not from chain {{chainName}}"
   },
   "connections": {
     "switchWalletTo": "Switch Wallet to {{chainName}}",
