+++ apps/davi/src/components/CustomRPC/CustomRPC.tsx
@@ -1,10 +1,6 @@
-import { getChainIcon } from 'utils';
 import { useNetwork } from 'wagmi';
-import { Chain, CustomRPCInput } from './components';
-import {
-  ChainOptionContainer as ChainContainer,
-  CustomRPCModal,
-} from './CustomRPC.styled';
+import { CustomRPCOption } from './components';
+import { CustomRPCModal } from './CustomRPC.styled';
 
 export const CustomRPC: React.FC = () => {
   const { chains } = useNetwork();
@@ -12,10 +8,7 @@ export const CustomRPC: React.FC = () => {
   return (
     <CustomRPCModal>
       {chains.map(chain => (
-        <ChainContainer key={chain.id}>
-          <Chain name={chain.name} icon={getChainIcon(chain.id)} />
-          <CustomRPCInput chain={chain} />
-        </ChainContainer>
+        <CustomRPCOption key={chain.id} chain={chain} />
       ))}
     </CustomRPCModal>
   );
